Finished Part 1

+++ Floater.pde
@@ -8,15 +8,17 @@ abstract class Floater //Do NOT modify the Floater class! Make changes in the Sp
   protected double myDirectionX, myDirectionY; //holds x and y coordinates of the vector for direction of travel   
   protected double myPointDirection; //holds current direction the ship is pointing in degrees    
   abstract public void setX(int x);  
-  abstract public int getX();   
+  abstract public double getX();   
   abstract public void setY(int y);   
-  abstract public int getY();   
+  abstract public double getY();   
   abstract public void setDirectionX(double x);   
   abstract public double getDirectionX();   
   abstract public void setDirectionY(double y);   
   abstract public double getDirectionY();   
   abstract public void setPointDirection(int degrees);   
   abstract public double getPointDirection(); 
+  //abstract public void setAccelerating(boolean x);
+  //abstract public boolean getAccelerating();
 
   //Accelerates the floater in the direction it is pointing (myPointDirection)   
   public void accelerate (double dAmount)
